+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## [0.3.0] Bugfixing
+- Fix behaviour when there's no bot token provided
+- Fix webhook URL parser
+
 ## [0.2.0] Commands
 - Some internal tomfoolery was done
 - You can now define discord slashcommands that can do stuff
 - Allow to differentiate between players that left and players that failed to connect
 
 ## [0.1.0] First release
-- it works? I think?
+- It works? I think?